requirements.txt
streamlit
anthropic
python-docx
gspread
google-auth
python-dotenv
fastapi
uvicorn
